+++ test/chex/board_test.exs
@@ -3,7 +3,9 @@ defmodule Chex.BoardTest do
   import AssertValue
   doctest Chex.Board
 
-  alias Chex.Board
+  alias Chex.{Board, Game}
+
+  @empty_board "8/8/8/4N3/8/8/8/8 w - - 0 1"
 
   test "file_offset" do
     results =
@@ -23,5 +25,31 @@ defmodule Chex.BoardTest do
       end
 
     assert_value(results == [true, false, false, true])
+  end
+
+  describe "Board.find_piece/2" do
+    test "returns the square of the first matching piece" do
+      {:ok, game} = Game.new()
+      assert {:a, 2} == Board.find_piece(game.board, {:pawn, :white})
+    end
+
+    test "returns nil when the piece is not on the board" do
+      {:ok, game} = Game.new(@empty_board)
+      assert nil == Board.find_piece(game.board, {:pawn, :white})
+    end
+  end
+
+  describe "Board.find_pieces/2" do
+    test "returns squares of all matching pieces" do
+      {:ok, game} = Game.new()
+
+      assert [h: 2, g: 2, f: 2, e: 2, d: 2, c: 2, b: 2, a: 2] ==
+               Board.find_pieces(game.board, {:pawn, :white})
+    end
+
+    test "returns an empty list when the piece is not on the board" do
+      {:ok, game} = Game.new(@empty_board)
+      assert [] == Board.find_pieces(game.board, {:pawn, :white})
+    end
   end
 end