+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Piece.trim/1` to remove starting square from a three element tuple containing
   a piece name, color, and starting square.
 - `Queen`, `Bishop`, `Knight`, and `Rook` modules.
+- `Game.find_piece/2` and `Game.find_pieces/2`.
 
 ### Changed